requirements.txt
# Core Dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document Processing
pypdf==3.17.0
pdfplumber==0.11.0
langchain==0.3.8
langchain-community==0.3.8

# Vector Store & Search
faiss-cpu==1.7.4
chromadb==0.5.3
rank-bm25==0.2.2
sentence-transformers==3.0.0

# LLM & AI
openai==1.60.1
transformers==4.44.2
torch==2.7.1
tiktoken==0.8.0

# Caching & Database
redis==5.0.3
sqlalchemy
asyncpg==0.30.0

# Monitoring & Testing
prometheus-client==0.21.0
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.2

# Utils
python-dotenv==1.0.1
pydantic==2.8.2
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.5
